agent: allow starting the agent without installing a signal handler
If the application the agent is being installed on also need to run some
shutdown code during an Interrupt signal, it's important to have the
ability to prevent gops agent from calling os.Exit(). This also adds the
requirement of exposing a method to manually cleanup garbage left by the
gops agent.

To allow custom agent flags now and in the future an Agent type was
created with a public boolean field HandleSignals.